refactor(utils): move utils into src - move utils folder into src - update import paths
