+++ src/utils/validator.js
@@ -1,4 +1,4 @@
-import { ERROR_MESSAGES } from '../src/constants/errors.js';
+import { ERROR_MESSAGES } from '../constants/errors.js';
 
 class Validator {
   static validateCarNames(names) {
@@ -35,13 +35,13 @@ class Validator {
   static validateGameCount(input) {
     const gameCount = Number(input);
 
-    // 입력받은 횟수가 숫자인지 확인
-    if (isNaN(gameCount)) {
-      throw new Error(ERROR_MESSAGES.INVALID_GAME_COUNT_TYPE);
+    // 입력받은 횟수가 정수인지 확인
+    if (!Number.isInteger(gameCount)) {
+      throw new Error(ERROR_MESSAGES.INVALID_GAME_COUNT_INTEGER);
     }
 
     // 입력받은 횟수가 0 이상인지 확인
-    if (gameCount < 0) {
+    if (gameCount <= 0) {
       throw new Error(ERROR_MESSAGES.INVALID_GAME_COUNT);
     }
   }
